Make "All variants" story a dev story
