+++ packages/react/src/SelectPanel/SelectPanel.features.stories.tsx
@@ -831,45 +831,3 @@ export const SingleSelectModal = () => {
     />
   )
 }
-
-export const AllVariants = () => {
-  return (
-    <>
-      <Text fontSize={3} fontWeight="bold">
-        Showcase of all the SelectPanel variants
-      </Text>
-      <br />
-      <Text>
-        Test the different interactions below to see how the SelectPanel behaves in different selection and anchoring
-        modes. The size of the screen also affects how the user interacts with the SelectPanel.
-      </Text>
-      <br />
-      <br />
-
-      <Text fontWeight="bold">Single Select Panel</Text>
-      <br />
-      <Text>This panel allows selecting a single item from the list.</Text>
-      <SingleSelect />
-      <br />
-
-      <Text fontWeight="bold">Single Select Modal</Text>
-      <br />
-      <Text>This modal allows selecting a single item with a modal interface.</Text>
-      <SingleSelectModal />
-      <br />
-
-      <Text fontWeight="bold">Multi Select Panel</Text>
-      <br />
-      <Text>This panel allows selecting multiple items from the list.</Text>
-      <WithOnCancel />
-      <br />
-
-      <Text fontWeight="bold">Multi Select Modal</Text>
-      <Text>
-        <br />
-        This modal allows selecting multiple items with a modal interface.
-      </Text>
-      <MultiSelectModal />
-    </>
-  )
-}
